backend/requirements.txt
langchain
langgraph
langchain-groq
langchain_community
agentipy
streamlit
python-dotenv